[clean]　カスタムフック化等のファイル整理
