+++ src/components/molecules/SearchInput.tsx
@@ -1,6 +1,7 @@
 import { Flex, Input } from "@chakra-ui/react";
-import { PrimaryButton } from "components/atoms/button/PrimaryButton";
 import { memo, MouseEvent, VFC } from "react";
+
+import { PrimaryButton } from "components/atoms/button/PrimaryButton";
 
 type Props = {
   value: string;
@@ -14,7 +15,7 @@ export const SearchInput: VFC<Props> = memo((props) => {
   const { value, onChange, submit, disabled, placeholder } = props;
   return (
     <Flex>
-      <Input value={value} onChange={onChange} placeholder={placeholder} p={4}/>
+      <Input value={value} onChange={onChange} placeholder={placeholder} p={4} />
       <PrimaryButton submit={submit} disabled={disabled} bg="green.100">
         検索
       </PrimaryButton>
