+++ src/components/organisms/detailPharmacy/ReportInfo.tsx
@@ -26,7 +26,7 @@ export const ReportInfo: VFC<Props> = memo((props) => {
                 <Th fontSize={{ base: "md", md: "lg" }}>届出基準名</Th>
                 <Th fontSize={{ base: "md", md: "lg" }}>点数</Th>
                 <Th fontSize={{ base: "md", md: "lg" }} textAlign="center">
-                  算定予想
+                  調剤基本料
                 </Th>
               </Tr>
             </Thead>
@@ -40,7 +40,7 @@ export const ReportInfo: VFC<Props> = memo((props) => {
           * 経過措置の都合上、地域支援体制加算が２つ存在する可能性がありますが、令和５年３月３１日までは、基本的に番号の若いものが算定されます。
         </Text>
         <Text fontSize={{ base: "12px", md: "15px" }}>
-          * 新型コロナウイルスによる臨時的な取り扱い等の都合により、点数が表記のものより高くなる可能性があります。
+          * 関連通知により、点数が表記のものより高くなる可能性があります。
         </Text>
         <Text fontSize={{ base: "12px", md: "15px" }}>
           * ３ヶ月以内に処方箋を持参かつ手帳を持参している場合は、４５ 点。その他の場合は、５９点算定がされます（例外あり）。
