+++ src/lib/api/pharmacy.ts
@@ -10,4 +10,4 @@ export const pharmacyReportImoprt = (formData: any) => client.post("pharmacies/p
 
 export const reportImoprt = (formData: any) => client.post("reports/report_import", formData);
 
-export const show = (id: number) => client.get<Array<Report>>(`pharmacies/${id}`);
+export const show = (id: any) => client.get<Array<Report>>(`pharmacies/${id}`);
